logging/pyproject.toml
[project]
name = "logging"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-monitor-opentelemetry>=1.7.0",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    "streamlit>=1.47.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "python-dotenv>=1.1.1",
]
